Feat: Update packaging methods

+++ src/commands/init.ts
@@ -7,13 +7,10 @@ import { Command } from "commander";
 
 import { handleError } from "@/utils/handle-error";
 import { logger } from "@/utils/logger";
-import {
-  getPackageManager,
-  checkPackageExists,
-} from "@/utils/get-package-info";
+import { getPackageManager, checkPackageExists } from "@/utils/get-package-info";
 import { resolveConfigPaths, Config } from "@/utils/get-config";
 
-import { getThemes, getInitData } from "@/utils/api";
+import { getInitData } from "@/utils/api";
 
 export const init = new Command()
   .name("init")
@@ -35,34 +32,27 @@ export const init = new Command()
         process.exit(1);
       }
 
-      const spinner = ora(`Validating project...`).start();
+      const spinner = ora(`Valodating project...`).start();
       // check for package.json
       if (!existsSync(path.resolve(cwd, "package.json"))) {
         spinner.fail();
-        logger.error(
-          `No package.json found in ${cwd}. Please run "npm init" first.`
-        );
+        logger.error(`No package.json found in ${cwd}. Please run "npm init" first.`);
         process.exit(1);
       }
 
       // check if either next or react exists
-      if (
-        !checkPackageExists("react", cwd) &&
-        !checkPackageExists("next", cwd)
-      ) {
+      if (!checkPackageExists("react", cwd) && !checkPackageExists("next", cwd)) {
         spinner.fail();
-        logger.error(
-          `Missing required dependencies. Please install React or Next first.`
-        );
+        logger.error(`Missing required dependencies. Please install React or Next first.`);
         process.exit(1);
       }
 
       spinner.succeed();
 
-      const projectConfig = await promptForProjectDetails(cwd);
+      const { resolvedConfig: projectConfig, overwrite } = await promptForProjectDetails(cwd);
       logger.info("Project configuration saved to atomic.components.json");
 
-      await runInit(cwd, projectConfig);
+      await runInit(cwd, projectConfig, overwrite);
     } catch (error) {
       handleError(error);
       // delete atomic.components.json if it exists
@@ -74,24 +64,19 @@ export const init = new Command()
   });
 
 async function promptForProjectDetails(cwd: string) {
-  const themes = await getThemes();
-
   const options = await prompts(
     [
-      {
-        type: "select",
-        name: "theme",
-        message: `Which theme would you like to use?`,
-        choices: themes.map((theme) => ({
-          title: theme.name,
-          value: theme.id,
-        })),
-      },
       {
         type: "text",
         name: "globalCSS",
         message: `Where is your global CSS file?`,
         initial: "app/globals.css",
+      },
+      {
+        type: "confirm",
+        name: "overwrite",
+        message: `Want to overwrite your global CSS?`,
+        initial: true,
       },
       {
         type: "text",
@@ -109,7 +94,6 @@ async function promptForProjectDetails(cwd: string) {
 
   const atomicComponentsConfig = {
     theme: {
-      name: options.theme,
       css: options.globalCSS,
     },
     alias: {
@@ -117,18 +101,10 @@ async function promptForProjectDetails(cwd: string) {
     },
   };
 
-  const resolvedConfig = await resolveConfigPaths(
-    cwd,
-    atomicComponentsConfig as Config
-  );
+  const resolvedConfig = await resolveConfigPaths(cwd, atomicComponentsConfig as Config);
 
-  if (
-    resolvedConfig.resolvedPaths.components === "" ||
-    resolvedConfig.resolvedPaths.css === ""
-  ) {
-    throw new Error(
-      `Failed to resolve components alias, please enter a valid alias.`
-    );
+  if (resolvedConfig.resolvedPaths.components === "" || resolvedConfig.resolvedPaths.css === "") {
+    throw new Error(`Failed to resolve components alias, please enter a valid alias.`);
   }
 
   const { proceed } = await prompts({
@@ -144,32 +120,29 @@ async function promptForProjectDetails(cwd: string) {
 
   const spinner = ora(`Writing atomic.components.json...`).start();
   const targetPath = path.resolve(cwd, "atomic.components.json");
-  await fs.writeFile(
-    targetPath,
-    JSON.stringify(atomicComponentsConfig, null, 2),
-    "utf8"
-  );
+  await fs.writeFile(targetPath, JSON.stringify(atomicComponentsConfig, null, 2), "utf8");
   spinner.succeed();
 
-  return resolvedConfig;
+  return { resolvedConfig, overwrite: options.overwrite };
 }
 
-export async function runInit(cwd: string, config: Config) {
+export async function runInit(cwd: string, config: Config, overwrite: boolean) {
   const spinner = ora(`Initializing project...`).start();
 
   if (!existsSync(config.resolvedPaths.components)) {
     await fs.mkdir(config.resolvedPaths.components, { recursive: true });
   }
 
-  const initData = await getInitData({
-    theme: config.theme.name,
-  });
+  const initData = await getInitData();
 
   // Write global CSS file. Create path if it doesn't exist.
   if (!existsSync(config.resolvedPaths.css)) {
     await fs.mkdir(path.dirname(config.resolvedPaths.css), { recursive: true });
   }
-  await fs.writeFile(config.resolvedPaths.css, initData.css, "utf8");
+  const { size } = await fs.stat(path.dirname(config.resolvedPaths.css));
+
+  if ((size !== 0 && overwrite) || size === 0)
+    await fs.writeFile(config.resolvedPaths.css, initData.css, "utf8");
 
   spinner.succeed();
